modified server test cases
