+++ network/testing.go
@@ -1,14 +1,22 @@
 package network
 
 import (
+	"fmt"
+	"io/ioutil"
+	"math/rand"
+	"os"
 	"sync/atomic"
 	"testing"
 	"time"
 
 	"github.com/0xPolygon/polygon-sdk/chain"
+	"github.com/0xPolygon/polygon-sdk/helper/common"
 	"github.com/0xPolygon/polygon-sdk/secrets"
 	"github.com/0xPolygon/polygon-sdk/secrets/local"
 	"github.com/hashicorp/go-hclog"
+	"github.com/libp2p/go-libp2p-core/crypto"
+	"github.com/libp2p/go-libp2p-core/peer"
+	"github.com/multiformats/go-multiaddr"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -108,4 +116,55 @@ func getTestConfig(callback func(c *Config)) *Config {
 	cfg.SecretsManager = secretsManager
 
 	return cfg
+}
+func GenerateTestMultiAddr(t *testing.T) multiaddr.Multiaddr {
+	libp2pKey, _, keyErr := GenerateAndEncodeLibp2pKey()
+	if keyErr != nil {
+		t.Fatalf("unable to generate libp2p key, %v", keyErr)
+	}
+	nodeId, err := peer.IDFromPrivateKey(libp2pKey)
+	assert.NoError(t, err)
+	rand.Seed(time.Now().Unix())
+	randomPort := rand.Intn(10) + 10010
+	addr, err := multiaddr.NewMultiaddr(fmt.Sprintf("/ip4/127.0.0.1/tcp/%d/p2p/%s", randomPort, nodeId))
+	assert.NoError(t, err)
+	return addr
+}
+func GenerateTestLibp2pKey(t *testing.T) (crypto.PrivKey, string) {
+	t.Helper()
+
+	dir, err := ioutil.TempDir(os.TempDir(), "")
+	assert.NoError(t, err)
+
+	// Instantiate the correct folder structure
+	setupErr := common.SetupDataDir(dir, []string{"libp2p"})
+	if setupErr != nil {
+		t.Fatalf("unable to generate libp2p folder structure, %v", setupErr)
+	}
+
+	localSecretsManager, factoryErr := local.SecretsManagerFactory(
+		nil,
+		&secrets.SecretsManagerParams{
+			Logger: hclog.NewNullLogger(),
+			Extra: map[string]interface{}{
+				secrets.Path: dir,
+			},
+		})
+	assert.NoError(t, factoryErr)
+
+	libp2pKey, libp2pKeyEncoded, keyErr := GenerateAndEncodeLibp2pKey()
+	if keyErr != nil {
+		t.Fatalf("unable to generate libp2p key, %v", keyErr)
+	}
+
+	if setErr := localSecretsManager.SetSecret(secrets.NetworkKey, libp2pKeyEncoded); setErr != nil {
+		t.Fatalf("unable to save libp2p key, %v", setErr)
+	}
+
+	t.Cleanup(func() {
+		// remove directory after test is done
+		assert.NoError(t, os.RemoveAll(dir))
+	})
+
+	return libp2pKey, dir
 }
